refactor(server): separate express app and https server for layered architecture
