+++ backend/.env.example
@@ -1,7 +1,20 @@
-# .env
-PORT=3000
-REDIRECT_URI=
+# Server Settings
+PORT=4000
 
-# .env.local
-CLIENT_ID=
-CLIENT_SECRET=
+# SoundCloud API
+SOUNDCLOUD_API_BASE_URL=https://api.soundcloud.com
+CLIENT_ID=your_client_id_here
+CLIENT_SECRET=your_client_secret_here
+REDIRECT_URI=https://localhost:3000/callback
+
+# MySQL
+MYSQL_HOST=localhost
+MYSQL_USER=root
+MYSQL_PASSWORD=your_mysql_password
+MYSQL_NAME=digbeats
+MYSQL_PORT=3306
+
+# Redis
+REDIS_HOST=localhost
+REDIS_PORT=6379
+
