--- backend/src/domain/entities/user.ts
@@ -1,8 +1,0 @@
-import { UserInfo } from "../valueObjects/userInfo";
-
-export class User {
-  constructor(
-    private readonly _userId: number,
-    private readonly _userInfo: UserInfo
-  ) {}
-}
