+++ backend/src/presentation/di/artistController.di.ts
@@ -1,6 +1,6 @@
 import { ArtistController } from "../controller/artistController";
 import { SearchArtistsUseCase } from "../../application/usecase/searchArtistsUseCase";
-import { TokenApplicationService } from "../../application/applicationSercices/tokenApplicationService";
+import { TokenApplicationService } from "../../application/applicationServices/tokenApplicationService";
 import { SessionRedisRepository } from "../../infrastructure/redis/sessionRedisRepository";
 import { TokenSoundCloudRepository } from "../../infrastructure/api/tokenSoundCloudRepository";
 import { ArtistSoundCloudRepository } from "../../infrastructure/api/artistSoundCloudRepository";
