+++ backend/src/presentation/di/userController.di.ts
@@ -1,6 +1,6 @@
 import { UserController } from "../controller/userController";
 import { FetchMyFollowingsUseCase } from "../../application/usecase/fetchMyFollowingsUseCase";
-import { TokenApplicationService } from "../../application/applicationSercices/tokenApplicationService";
+import { TokenApplicationService } from "../../application/applicationServices/tokenApplicationService";
 import { SessionRedisRepository } from "../../infrastructure/redis/sessionRedisRepository";
 import { TokenSoundCloudRepository } from "../../infrastructure/api/tokenSoundCloudRepository";
 import { UserSoundCloudRepository } from "../../infrastructure/api/userSoundCloudRepository";
